# internal/asset/assets/openapi.yaml
...
openapi: "3.1.2"
info:
    title: "Keys"
    description: Keys provides HTTP endpoints that can be called directly when browser or keyboard input isn't convenient.
    version: "1.0.0"
servers:
    - url: "{{ .PublicUrl }}"
...
paths:
...
    /trigger/{key}:
        post:
            summary: Press one or more keys
            description: The equivalent to pressing a key on the keyboard.
            tags:
                - trigger
            operationId: trigger
            parameters:
                - name: key
                  in: path
                  required: true
                  description: The physical keys to be pressed.
                  schema:
                      type: string
                      example: h
            responses:
                "200":
                    description: Stdout of the command associated with the specified key.
                    headers:
                        X-Keys-State:
                            description: If the pressed key toggles between multiple commands, its current state.
                            schema:
                                type: string
                        X-Keys-Locked:
                            description: |
                                If the pressed key's command was "lock" or "unlock",
                                the current keyboard lock state as a boolean integer. A user-facing
                                confirmation message is always produced by these commands,
                                so they are never a 204.
                            schema:
                                type: integer
                    content:
                        text/plain:
                            schema:
                                type: string
                                example: hello world
                        text/html:
                            schema:
                                type: string
                "204":
                    description: |
                        Successful invocation of a key whose command produced no output,
                        or was configured to suppress output.
                    headers:
                        X-Keys-State:
                            description: Same as for 200 response.
                            schema:
                                type: string
                "405":
                    description: Unknown key.
...